refactor(ui): extract queryClient into dedicated lib/queryClient.ts
Keeps lib/utils.ts as a pure helper module (only cn()) so that
components importing cn do not pull in @tanstack/react-query
or instantiate QueryClient as a side effect.
