+++ ui/src/lib/utils.ts
@@ -1,15 +1,5 @@
-import { QueryClient } from '@tanstack/react-query'
 import { type ClassValue, clsx } from 'clsx'
 import { twMerge } from 'tailwind-merge'
-
-export const queryClient = new QueryClient({
-  defaultOptions: {
-    queries: {
-      staleTime: 30_000,
-      retry: 1,
-    },
-  },
-})
 
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs))
