+++ ui/src/main.tsx
@@ -1,15 +1,15 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
-import { QueryClientProvider } from '@tanstack/react-query'
-import App from './App'
-import { queryClient } from './lib/utils'
-import './index.css'
-import './App.css'
+import React from "react";
+import ReactDOM from "react-dom/client";
+import { QueryClientProvider } from "@tanstack/react-query";
+import App from "./App";
+import { queryClient } from "./lib/queryClient";
+import "./index.css";
+import "./App.css";
 
-ReactDOM.createRoot(document.getElementById('root')!).render(
+ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <QueryClientProvider client={queryClient}>
       <App />
     </QueryClientProvider>
-  </React.StrictMode>
-)
+  </React.StrictMode>,
+);
